# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AuthData:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  handlers.AuthResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.AuthData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.ConversationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.ConversationResponse:
    properties:
      data:
        $ref: '#/definitions/models.Conversation'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.LoginRequest:
    properties:
      identifier:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  handlers.MessageResp:
    properties:
      data:
        $ref: '#/definitions/models.Message'
      success:
        example: true
        type: boolean
    type: object
  handlers.MessageResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      success:
        example: true
        type: boolean
    type: object
...
  handlers.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: secret123
        type: string
      phone_number:
        example: "+237600000000"
        type: string
    type: object
...
  handlers.SendMessageRequest:
    properties:
      content:
        example: Hello, I am interested in renting this property.
        type: string
    type: object
  handlers.StartConversationRequest:
    properties:
      message:
        example: Is this property still available?
        type: string
      property_id:
        example: 42
        type: integer
      recipient_id:
        example: 5
        type: integer
    type: object
...
  handlers.UserResponse:
    properties:
      data:
        $ref: '#/definitions/models.User'
      success:
        example: true
        type: boolean
    type: object
...
  models.Conversation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      initiator:
        $ref: '#/definitions/models.User'
      initiator_id:
        type: integer
      last_message:
        $ref: '#/definitions/models.Message'
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      property:
        $ref: '#/definitions/models.Property'
      property_id:
        type: integer
      recipient:
        $ref: '#/definitions/models.User'
      recipient_id:
        type: integer
      unread_count:
        type: integer
      updated_at:
        type: string
    type: object
  models.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      sender:
        $ref: '#/definitions/models.User'
      sender_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Property:
    properties:
      average_rating:
        description: Computed (not stored)
        type: number
      bathrooms:
        type: integer
      city:
        type: string
      country:
        description: Location — from broad to precise
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      exact_address:
        type: string
      furnished:
        description: Amenities
        type: boolean
      has_courtyard:
        type: boolean
      has_electricity:
        type: boolean
      has_water:
        type: boolean
      has_wifi:
        type: boolean
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.PropertyImage'
        type: array
      is_available:
        type: boolean
      is_favorited:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        description: sub-area within the city (optional)
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      owner_id:
        type: integer
      price:
        description: Pricing — ISO 4217 currency codes (KES, USD, EUR, GBP, NGN, GHS,
          TZS, UGX …)
        type: number
      property_type:
        description: Bedsitter, Studio, Apartment, Maisonette, Bungalow, Townhouse,
          Villa, Commercial
        type: string
      review_count:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      rooms:
        description: Size & layout
        type: integer
      shower_type:
        description: en_suite, shared
        type: string
      state:
        description: state / county / region / province (optional)
        type: string
      surface:
        description: m²
        type: number
      title:
        description: Listing info
        type: string
      transaction_type:
        description: rent, sale
        type: string
      updated_at:
        type: string
    type: object
  models.PropertyImage:
    properties:
      created_at:
        type: string
      filename:
        type: string
      id:
        type: integer
      is_main:
        type: boolean
      property_id:
        type: integer
    type: object
...
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      property_id:
        type: integer
      rating:
        description: 1-5
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      locale:
        type: string
      member_since:
        type: string
      phone_number:
        type: string
      profile_picture:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gidana.com
    name: Gidana Support
  description: Real-estate platform API — property listings, rentals, messaging, digital
    wallets & payments.
  title: Gidana API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login with email/phone and password
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /conversations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ConversationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all conversations for the current user
      tags:
      - conversations
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.StartConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ConversationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a new conversation with another user
      tags:
      - conversations
  /conversations/{id}:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ConversationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a single conversation with its messages
      tags:
      - conversations
  /conversations/{id}/messages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message content
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SendMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send a message in a conversation
      tags:
      - conversations
  /conversations/{id}/messages/{msgID}:
    delete:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message ID
        in: path
        name: msgID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a message
      tags:
      - conversations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"